+++ Sources/MapboxNavigation/CarPlayManager.swift
@@ -10,7 +10,6 @@ import MapboxMaps
  
  - note: It is very important you have a single `CarPlayManager` instance at any given time. This should be managed by your `UIApplicationDelegate` class if you choose to supply your `accessToken` to the `CarPlayManager.eventsManager` via `NavigationEventsManager` initializer, instead of the Info.plist.
  */
-@available(iOS 12.0, *)
 public class CarPlayManager: NSObject {
     
     // MARK: CarPlay Infrastructure
@@ -91,6 +90,7 @@ public class CarPlayManager: NSObject {
 
     private weak var navigationService: NavigationService?
     private var idleTimerCancellable: IdleTimerManager.Cancellable?
+    private var indexedRouteResponse: IndexedRouteResponse?
     
     /**
      Programatically begins a CarPlay turn-by-turn navigation session.
@@ -371,7 +371,6 @@ public class CarPlayManager: NSObject {
 
 // MARK: CPApplicationDelegate Methods
 
-@available(iOS 12.0, *)
 extension CarPlayManager: CPApplicationDelegate {
     
     public func application(_ application: UIApplication,
@@ -496,7 +495,6 @@ extension CarPlayManager: CPApplicationDelegate {
 
 // MARK: CPInterfaceControllerDelegate Methods
 
-@available(iOS 12.0, *)
 extension CarPlayManager: CPInterfaceControllerDelegate {
     
     public func templateWillAppear(_ template: CPTemplate, animated: Bool) {
@@ -507,8 +505,7 @@ extension CarPlayManager: CPInterfaceControllerDelegate {
             carPlayMapViewController.recenterButton.isHidden = true
         }
         
-        if let userInfo = template.userInfo as? Dictionary<String, Any>,
-           let currentActivity = userInfo[CarPlayManager.currentActivityKey] as? CarPlayActivity {
+        if let currentActivity = template.currentActivity {
             self.currentActivity = currentActivity
         } else {
             self.currentActivity = nil
@@ -519,13 +516,8 @@ extension CarPlayManager: CPInterfaceControllerDelegate {
         delegate?.carPlayManager(self, templateDidAppear: template, animated: animated)
         
         guard interfaceController?.topTemplate == mainMapTemplate,
-              template == interfaceController?.rootTemplate,
-              let carPlayMapViewController = carPlayMapViewController else { return }
-        
-        let navigationMapView = carPlayMapViewController.navigationMapView
-        navigationMapView.removeRoutes()
-        navigationMapView.removeContinuousAlternativesRoutes()
-        navigationMapView.removeWaypoints()
+              template == interfaceController?.rootTemplate else { return }
+        self.removeRoutesFromMap()
     }
     
     public func templateWillDisappear(_ template: CPTemplate, animated: Bool) {
@@ -535,7 +527,7 @@ extension CarPlayManager: CPInterfaceControllerDelegate {
               let topTemplate = interfaceController.topTemplate,
               type(of: topTemplate) == CPSearchTemplate.self ||
                 interfaceController.templates.count == 1 else { return }
-        
+
         navigationMapView?.navigationCamera.follow()
     }
     
@@ -544,7 +536,6 @@ extension CarPlayManager: CPInterfaceControllerDelegate {
     }
 }
 
-@available(iOS 12.0, *)
 extension CarPlayManager {
     
     // MARK: Route Preview
@@ -629,8 +620,25 @@ extension CarPlayManager {
      previewed.
      */
     public func previewRoutes(for indexedRouteResponse: IndexedRouteResponse) {
+        guard shouldPreviewRoutes(for: indexedRouteResponse) else { return }
         let trip = CPTrip(indexedRouteResponse: indexedRouteResponse)
         previewRoutes(for: trip)
+    }
+    
+    /**
+     Allows to cancel routes preview on CarPlay .
+     */
+    public func cancelRoutesPreview() {
+        guard self.indexedRouteResponse != nil else { return }
+        self.indexedRouteResponse = nil
+        mainMapTemplate?.hideTripPreviews()
+        popToRootTemplate(interfaceController: interfaceController, animated: true)
+        delegate?.carPlayManagerDidCancelPreview(self)
+    }
+    
+    func shouldPreviewRoutes(for indexedRouteResponse: IndexedRouteResponse) -> Bool {
+        guard self.indexedRouteResponse?.currentRoute == indexedRouteResponse.currentRoute else { return true }
+        return self.indexedRouteResponse?.routeResponse.routes != indexedRouteResponse.routeResponse.routes
     }
     
     func previewRoutes(for trip: CPTrip) {
@@ -650,8 +658,23 @@ extension CarPlayManager {
             previewText = customPreviewText
         }
         
+        previewMapTemplate.backButton = defaultTripPreviewBackButton()
         previewMapTemplate.showTripPreviews([modifiedTrip], textConfiguration: previewText)
-        interfaceController.pushTemplate(previewMapTemplate, animated: true)
+        
+        if currentActivity == .previewing {
+            interfaceController.popTemplate(animated: false)
+            interfaceController.pushTemplate(previewMapTemplate, animated: false)
+        } else {
+            interfaceController.pushTemplate(previewMapTemplate, animated: true)
+        }
+    }
+    
+    func removeRoutesFromMap() {
+        indexedRouteResponse = nil
+        guard let navigationMapView = carPlayMapViewController?.navigationMapView else { return }
+        navigationMapView.removeRoutes()
+        navigationMapView.removeContinuousAlternativesRoutes()
+        navigationMapView.removeWaypoints()
     }
     
     func calculate(_ options: RouteOptions, completionHandler: @escaping RoutingProvider.IndexedRouteResponseCompletionHandler) {
@@ -705,11 +728,22 @@ extension CarPlayManager {
                                                                 overviewButtonTitle: overviewTitle)
         return defaultPreviewText
     }
+    
+    private func defaultTripPreviewBackButton() -> CPBarButton {
+        let backButton = CPBarButton(type: .text) { [weak self] (button: CPBarButton) in
+            guard let self = self else { return }
+            self.cancelRoutesPreview()
+        }
+        backButton.title = NSLocalizedString("CARPLAY_PREVIEW_BACK",
+                                             bundle: .mapboxNavigation,
+                                             value: "BACK",
+                                             comment: "Title for trip preview back button")
+        return backButton
+    }
 }
 
 // MARK: CPMapTemplateDelegate Methods
 
-@available(iOS 12.0, *)
 extension CarPlayManager: CPMapTemplateDelegate {
     
     public func mapTemplate(_ mapTemplate: CPMapTemplate,
@@ -720,7 +754,7 @@ extension CarPlayManager: CPMapTemplateDelegate {
                   return
               }
         
-        guard let indexedRouteResponse = routeChoice.indexedRouteResponseUserInfo?.indexedRouteResponse else {
+        guard let indexedRouteResponse = routeChoice.indexedRouteResponse else {
             preconditionFailure("CPRouteChoice should contain `IndexedRouteResponseUserInfo` struct.")
         }
 
@@ -775,10 +809,7 @@ extension CarPlayManager: CPMapTemplateDelegate {
             self.delegate?.carPlayManager(self, didPresent: carPlayNavigationViewController)
         }
         
-        let navigationMapView = carPlayMapViewController.navigationMapView
-        navigationMapView.removeRoutes()
-        navigationMapView.removeContinuousAlternativesRoutes()
-        navigationMapView.removeWaypoints()
+        self.removeRoutesFromMap()
     }
 
     func navigationMapTemplate() -> CPMapTemplate {
@@ -800,7 +831,7 @@ extension CarPlayManager: CPMapTemplateDelegate {
                             using routeChoice: CPRouteChoice) {
         guard let carPlayMapViewController = carPlayMapViewController else { return }
         
-        guard let indexedRouteResponse = routeChoice.indexedRouteResponseUserInfo?.indexedRouteResponse,
+        guard let indexedRouteResponse = routeChoice.indexedRouteResponse,
               let route = indexedRouteResponse.currentRoute,
               var routes = indexedRouteResponse.routeResponse.routes else {
                   preconditionFailure("CPRouteChoice should contain `IndexedRouteResponseUserInfo` struct.")
@@ -819,7 +850,7 @@ extension CarPlayManager: CPMapTemplateDelegate {
         navigationMapView.showcase(routes,
                                    routesPresentationStyle: .all(shouldFit: true, cameraOptions: cameraOptions),
                                    animated: true)
-        
+        self.indexedRouteResponse = indexedRouteResponse
         delegate?.carPlayManager(self, selectedPreviewFor: trip, using: routeChoice)
     }
 
@@ -828,9 +859,7 @@ extension CarPlayManager: CPMapTemplateDelegate {
             return
         }
         let navigationMapView = carPlayMapViewController.navigationMapView
-        navigationMapView.removeRoutes()
-        navigationMapView.removeContinuousAlternativesRoutes()
-        navigationMapView.removeWaypoints()
+        self.removeRoutesFromMap()
         if let passiveLocationProvider = navigationMapView.mapView.location.locationProvider as? PassiveLocationProvider {
             passiveLocationProvider.locationManager.resumeTripSession()
             carPlayMapViewController.subscribeForFreeDriveNotifications()
@@ -903,10 +932,7 @@ extension CarPlayManager: CPMapTemplateDelegate {
     }
     
     public func mapTemplateDidDismissPanningInterface(_ mapTemplate: CPMapTemplate) {
-        guard let userInfo = mapTemplate.userInfo as? CarPlayUserInfo,
-              let currentActivity = userInfo[CarPlayManager.currentActivityKey] as? CarPlayActivity else {
-                  return
-              }
+        guard let currentActivity = mapTemplate.currentActivity else { return }
         
         self.currentActivity = currentActivity
         
@@ -1005,10 +1031,7 @@ extension CarPlayManager: CPMapTemplateDelegate {
      - parameter mapTemplate: `CPMapTemplate` instance, for which buttons update will be performed.
      */
     private func updateNavigationButtons(for mapTemplate: CPMapTemplate) {
-        guard let userInfo = mapTemplate.userInfo as? CarPlayUserInfo,
-              let currentActivity = userInfo[CarPlayManager.currentActivityKey] as? CarPlayActivity else {
-                  return
-              }
+        guard let currentActivity = mapTemplate.currentActivity else { return }
 
         let traitCollection: UITraitCollection
         if let carPlayNavigationViewController = carPlayNavigationViewController {
@@ -1074,7 +1097,6 @@ extension CarPlayManager: CPMapTemplateDelegate {
 
 // MARK: CarPlayNavigationViewControllerDelegate Methods
 
-@available(iOS 12.0, *)
 extension CarPlayManager: CarPlayNavigationViewControllerDelegate {
     
     public func carPlayNavigationViewControllerWillDismiss(_ carPlayNavigationViewController: CarPlayNavigationViewController,
@@ -1161,7 +1183,6 @@ extension CarPlayManager: CarPlayNavigationViewControllerDelegate {
 
 // MARK: CarPlayMapViewControllerDelegate Methods
 
-@available(iOS 12.0, *)
 extension CarPlayManager: CarPlayMapViewControllerDelegate {
     
     public func carPlayMapViewController(_ carPlayMapViewController: CarPlayMapViewController,
@@ -1210,7 +1231,6 @@ extension CarPlayManager: CarPlayMapViewControllerDelegate {
 
 // MARK: MapTemplateProviderDelegate Methods
 
-@available(iOS 12.0, *)
 extension CarPlayManager: MapTemplateProviderDelegate {
     
     func mapTemplateProvider(_ provider: MapTemplateProvider,
@@ -1283,12 +1303,12 @@ extension CarPlayManager {
         idleTimerCancellable = nil
         
         unsubscribeFromNotifications()
+        indexedRouteResponse = nil
     }
 }
 
 // MARK: CarPlayManager Constants
 
-@available(iOS 12.0, *)
 extension CarPlayManager {
     
     static let currentActivityKey = "com.mapbox.navigation.currentActivity"