+++ Sources/MapboxNavigation/CarPlayMapViewControllerDelegate.swift
@@ -4,7 +4,6 @@ import MapboxCoreNavigation
 /**
  The `CarPlayMapViewControllerDelegate` protocol provides methods for reacting to events during free-drive navigation or route previewing in `CarPlayMapViewController`.
  */
-@available(iOS 12.0, *)
 public protocol CarPlayMapViewControllerDelegate: AnyObject, UnimplementedLogging {
     
     /**
@@ -71,7 +70,6 @@ public protocol CarPlayMapViewControllerDelegate: AnyObject, UnimplementedLoggin
                                   willAdd layer: Layer) -> Layer?
 }
 
-@available(iOS 12.0, *)
 public extension CarPlayMapViewControllerDelegate {
     
     /**