+++ Sources/MapboxNavigation/CongestionLevel+CarPlay.swift
@@ -6,7 +6,6 @@ extension CongestionLevel {
     /**
      Converts a CongestionLevel to a CPTimeRemainingColor.
      */
-    @available(iOS 12.0, *)
     public var asCPTimeRemainingColor: CPTimeRemainingColor {
         switch self {
         case .unknown: